+++ src/adapters/index.ts
@@ -8,4 +8,6 @@ export { ContinueAdapter } from './ContinueAdapter';
 export { CopilotAdapter } from './CopilotAdapter';
 export { CursorAdapter } from './CursorAdapter';
 export { CustomAdapter } from './CustomAdapter';
+export type { PresetAdapterConfig, PresetFileType } from './PresetAdapter';
+export { PRESET_ADAPTERS, PresetAdapter } from './PresetAdapter';
 export { RulesAdapter } from './RulesAdapter';